requirements.txt
# RetailNova Enterprise Data Platform
# Python Dependencies
# =============================================
# Install:  pip install -r requirements.txt
#
# Linux notes:
#   - psycopg2-binary: prebuilt, no libpq-dev needed
#   - pyodbc: needs unixODBC headers (only for SQL Server from host)
#             sudo apt install unixodbc-dev   (Ubuntu/Debian)
#             sudo dnf install unixODBC-devel  (Fedora/RHEL)
#   - pyspark: downloads ~300MB on first install
#   - For unit tests: only pyspark + delta-spark + pytest are needed
# =============================================

# ── Core Spark ────────────────────────────────
pyspark==3.5.0
delta-spark==3.1.0

# ── Database connectivity ──────────────────────
psycopg2-binary>=2.9.9    # PostgreSQL (metadata) — no C headers needed
# pyodbc>=5.0.0            # SQL Server from host — OPTIONAL
#                          # Tests mock the DB, so pyodbc not needed for testing

# ── Object storage ────────────────────────────
boto3>=1.34.0             # MinIO / S3 Python client
minio>=7.2.0

# ── Data utilities ─────────────────────────────
pandas>=2.0.0
pyarrow>=14.0.0

# ── Testing ────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.2.0

# ── Utilities ──────────────────────────────────
python-dotenv>=1.0.0      # Load .env file
